Add loadTasklist function on startup
The chatbot has no ability to retain previously written tasks by
the user.

It would enhance the functionality of the chatbot to be able to
remember tasks so it needs to be able to load data.

Let's:
Use a text file to load Todo, Deadline, and Event items when the
chatbot starts, with the use of the java.nio.file library.
Ensure the chatbot creates a data folder and an empty text file
if they do not exist on startup.

If the chatbot is able to load data from a text file, we can alter
this text file to include user data.