Add convertToDataFormat function to Task and its children
We have no convenient way of producing the required data format for
a Task class and its children.

We need a function to streamline the process of getting the string
to be written to text file.

Let's include the function in Task and use polymorphism to create
the child's own convertToDataFormat function

Using polymorphism reduces the amount of checks needed for what
format should be used depending on the class.

There are common fields in the data between Task and its children
so polymorphism is suitable to be used.