Add overwriteTasklist and appendToTasklist functions
we have no functions to modify the data in the text file.

we need a way to save the user's changes to the tasklist to
complement the load tasklist functionality.

Lets,
*create a appendToTasklist function that append a new line of data
to the text file.
*create an overwriteTasklist function that loop through the tasklist
arraylist to overwrite the text file data by writing to it line by
line.

We use appendToTasklist when adding a new Task to the Text for
efficiency (since it directly adds to the list) while overwrite
is only used for modifying existing data (e.g. mark and delete).