Move processLocalDateTime method to Event and Deadline class
The processLocalDateTime method is located in main.

Placing processLocalDateTime in main caused repetitive calls
since whenever event or deadline object was created, String had
to be converted to LocalDateTime

Let's move the processLocalDateTime method into the Event and
Deadline class, as it directly pertains to parsing the dates
required by only these two classes.

Reduces the amount of times this function is called