Change method names with tasklist to taskList
Method names used tasklist but the class being referred to is
named TaskList.

This creates inconsistency in the naming convention.

Let's change all tasklist in method names to taskList.

This makes it more evident that our naming convention for method
names adheres to camelcase.

Note that junit test case methods are also affected by this change.