Extract methods from parse method
Every switch case statement for a command type has the logic of
processing that command contained in the switch case statement.

The lines of code for Parse exceeded 30 lines, making it hard to
understand and maintain.

Let's extract out a process method for each command type.

This improves readability and ensures that SLAP is easily
maintained in the switch case statements.