Extract methods from loadTaskListData method
loadTaskListData reads the text file data, parses the data into
tasks and adds the tasks to the task list. It also creates the data
file and directory if they does not exist.

The logic for operations are all inside of loadTaskListData, making
the lines of code over 30 and violating SLAP in certain areas.

Let's extract out these operations into methods. In particular, we
have:
*a method for creating the data file and directory if they does
not exist
*a method for splitting and trimming the data
*a method for creating and adding todo from the data
*a method for creating and adding deadline from the data
*a method for creating and adding event from the data

This improve code readability and ensures that it is easier to
maintain.