Add assertions to various methods
There is no usage of assertions in the code.

Assertions allow us to validate assumptions made during runtime,
ensuring the code behaves as expected.

Let's add assertions to our Command subclasses, Task subclasses and
exception class. In particular, we can perform assertions on the
fields in these classes to ensure they are not null or have
unexpected values during runtime.

This ensures that much fewer bugs go undetected as the program is
developed.